Eagerly compute object lifetime default
Simplifies the code a bit but might be detrimental to perf.
